requirements.txt
Pillow==9.0.0
requests==2.25.1
ruamel.base==1.0.0
